Extract Kompot into separate kernel library
