+++ src/kernel/vfs/CMakeLists.txt
@@ -2,3 +2,4 @@ SETUP_LIBRARY(vfs)
 
 target_link_libraries(vfs PRIVATE Debug)
 target_link_libraries(vfs PRIVATE stdbigos)
+target_link_libraries(vfs PRIVATE Kompot)
