+++ src/lib/kompot/CMakeLists.txt
@@ -1,0 +1,4 @@
+SETUP_LIBRARY(Kompot)
+
+target_link_libraries(Kompot PRIVATE Debug)
+target_link_libraries(Kompot PRIVATE stdbigos)
